requirements.txt
google.generativeai
serial
pyttsx3
speechRecognition
keyboard
pyserial
PyAudio